Week 6 Fri - modify Board, Comment

+++ board/src/main/java/mvc/command/DeleteBoardHandler.java
@@ -15,12 +15,18 @@ public class DeleteBoardHandler implements CommandHandler {
 			throws Exception {
 		try {
 			String noVal = req.getParameter("no");
+			String pageNoVal = req.getParameter("pageNo");
+			String pwd = req.getParameter("pwd");
 			int no = Integer.parseInt(noVal);
-			
-			Board board = boardService.getBoard(no, false);
-			boardService.delete(no);
-			req.setAttribute("board", board);
-			return "list.do";
+			int pageNo = Integer.parseInt(pageNoVal);
+			if(pwd == null) {
+				Board board = boardService.getBoard(no, false);
+				boardService.delete(no);
+				req.setAttribute("board", board);
+				return "list.do?pageNo=" + pageNo;
+			} else {
+				return "read.do?no=" + no + "&pageNo=" + pageNo;
+			}
 		} catch (BoardNotFoundException e) {
 			req.getServletContext().log("no board", e);
 			res.sendError(HttpServletResponse.SC_NOT_FOUND);
